api-master/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
email-validator>=2.1.0.post1 
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb
langchain-chroma
loguru>=0.7.3
stripe
supabase

# pip install -r requirements.txt